pyproject.toml
[tool.poetry]
name = "sleeper-data-pypline"
version = "0.1.0"
description = ""
authors = ["Parker Johnson <parkerjohnsonwebdev+gitconfig@gmail.com>"]
readme = "README.md"
packages = [{include = "sleeper_data_pypline"}]

[tool.poetry.dependencies]
python = "^3.11"
sleeper-api-wrapper = "^1.0.7"
pymongo = "3.11"
python-dotenv = "^1.0.0"
dnspython = {extras = ["srv"], version = "^2.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
